Add python script to create files of 5-letter words
Adds python script to request JSON file of all words. Filters the data
of only 5-letter words. Writes the 5-letter words to separate JSON
files as an array and object with key/value pairs. The array can be
used to loop over all words and the object with key/value pairs can
be used for O(1) read speed to check if a 5-letter word is valid.
Adds the JSON files with 5-letter words to be imported by the
TypeScript scripts.