Add logic script that checks for valid grid designs
Adds types and interfaces to help with type checking.
Adds global variable to hold array of all 5-letter words to be used in
any function in the logic script.
Adds and exports function that returns true if a 5-letter word is found
in valid list of 5-letter words.
Adds function that checks if a proposed grid design is possible to
achieve in Wordle.
Adds function that calculates the 'rarity' value of a grid design and
the number of guesses for each row.
Adds function that creates new grid design with certain color values
swapped.
Adds function that creates new grid design with color values flipped
horizontally to mirror a grid design.
Adds function that checks if a grid design is horizontally symmetric
to avoid mirroring a grid design that is already mirrored by being
symmetric.
Adds function that returns every variation of a base grid design. A
grid design can have each of their color values swapped and can also
be mirrored by flipping along the horizontal direction.
Adds, and exports as default, a function that returns an array of all
valid grid designs for the provided solution word.
Adds and exports function that returns a valid grid design with guesses
given a solution word and base grid design. Returns undefined if the
solution is not a possible grid design.
Adds function that returns an array of guesses to complete a single
row of a grid design.