Add component that finds valid grid designs
Adds component that checks what grid designs are valid with given
solution word.
User can submit word to be used as solution word and any valid grid
designs are found. Guesses for each valid grid design are provided when
the User clicks on a valid grid design to see additional details in a
modal.