Add component to display more detailed valid grid designs
Adds component that displays all details of a valid grid design,
including the guesses array for each row.
Props include a handler function to close the detailed modal whenever
the User clicks outside the modal or the close button on the modal
itself.
Adds ability for User to select a single guess from each row to highlight.