Add grid designs array parameter to checkGridDesigns function
Adds optional parameter to checkGridDesigns function of an array of grid
designs to check. If no argument is passed, the default array of grid
designs is used. This allows just one grid design to be tested inside
the WordleArtDesignCreator component so only that maximum 12 possible
variations of the design is checked whenever the User changes a node
color.