Add component for User to create a grid design
Adds componenent for User to create a grid design and check if it's
valid, or any variations valid, with given solution word.
User can create their own grid design by clicking on each node to toggle
through each of the 3 possible colors. Every change to the custom grid, or
a new solution word is submitted, the design is checked for validity. If the
User grid design is valid, they can view a detailed version showing the guesses
for each row.
Checks if any variations of the design are valid and displays them. Each can
be clicked to view more details in a modal, including the guesses.
Adds button for User to fill the whole grid of any of the 3 colors with one
button.
Adds placeholder button and submit handler to add custom grid design to
the library of grid designs checked in Checker component.